requirements.txt
Django>=5.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
django-rest-framework>=0.1.0
djangorestframework-simplejwt>=5.3.1
python-magic>=0.4.27
PyPDF2>=3.0.0
python-docx>=0.8.11
